Rename footer and input components
